dll: Provide 64-bit and 32-bit MSVC DLL library
DLL is a HIDAPI C library compiled with Visual Studio 2017 with fix
which enables sending/receiving feature requests on devices claimed by
Windows such as mouse or keyboard:
https://github.com/signal11/hidapi/pull/335/files

For 64-bit Python interpreter a 64-bit DLL library is needed. It is also
provided and will be loaded automatically if the program cannot load
32-bit library.

DLL files also include a fix for enumeration and opening of
non-composite HID devices:
https://github.com/signal11/hidapi/pull/418

DLLs are in "Release" version so they do not require Windows debug
libraries from user.

Signed-off-by: Kubicz, Filip <Filip.Kubicz@nordicsemi.no>